Prelogin page setup seccessfully . MockDB and User Class Added and Imports to main page so far. Successfull login page completed so far . Makes great templating object for future uses. Added functionality log user out. Added password hashing and password. Adding users to mock users to mockdbhelper class . Verifying that password matches up to password in db. Added User registration from form corrected. Successful registration and login created for user. Html templating has taken placing using jinja and extention is occurring. Added routed pages html to the application. HTML layouts added and improved.
Added Methods to the mockdbhelper class and more configuration was passed to the config file. A function was added to the main file as well. Function to remove table was added to page. Added Token to generate bitly link successfully.

Successfully enable feature to remove pending request and added meta refresh.Final touches added to application in terms of bringing on board a prod database and creating the classes for that db.